feat: Add Lambda Bot Authorizer and API Gateway
This commit introduces a new Lambda function called "LambdaBotAuthorizer".
The function verifies the authenticity of incoming requests by comparing
the received token header with the known secret token stored in AWS SSM.
A new route (/hook/event) is added to the API Gateway to allow incoming requests,
which are forwarded to the "TelegramBot" Lambda function after they pass the authorizer.
This update improves the security and accessibility of the bot by ensuring only
valid requests reach the bot and providing a HTTP(S) endpoint for the bot.

Additional changes include:
- The bank status updater function's schedule is changed from every 5 minutes to every 2 hours.
- The logging level for com.yuriytkach package is set to DEBUG for better traceability.
- A new application property file is added to manage secret name and telegram header.
- A SecretsReader interface and a SsmSecretsReader implementation are introduced for reading secrets